+++ SimPEG/electromagnetics/base_1d.py
@@ -1,39 +1,46 @@
-from SimPEG import maps, utils, props
-from SimPEG.simulation import BaseSimulation
-import numpy as np
-from .sources import *
-from .survey import BaseEM1DSurvey, EM1DSurveyFD, EM1DSurveyTD
-from .supporting_functions.kernels import *
-from scipy import sparse as sp
-from scipy.constants import mu_0
-from scipy.interpolate import InterpolatedUnivariateSpline as iuSpline
-from scipy.linalg import block_diag
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
 import properties
+from scipy.constants import mu_0
+import numpy as np
+from scipy import sparse as sp
+from scipy.linalg import block_diag
 
-from empymod.utils import check_time
-from empymod import filters
-from empymod.transform import dlf, fourier_dlf, get_dlf_points
-from empymod.utils import check_hankel
+from ..data import Data
+from ..maps import IdentityMap
+from ..simulation import BaseSimulation
+from ..survey import BaseSurvey, BaseSrc
+from .. import utils
+from ..utils import sdiag, Zero, mkvc
+from .. import props
 
-from .known_waveforms import (
-    piecewise_pulse_fast,
-    butterworth_type_filter, butter_lowpass_filter
-)
-
-try:
-    from multiprocessing import Pool
-    from sys import platform
-except ImportError:
-    print("multiprocessing is not available")
-    PARALLEL = False
-else:
-    PARALLEL = True
-    import multiprocessing
+__all__ = ["BaseEM1DSimulation", "BaseStitchedEM1DSimulation"]
 
 
-#######################################################################
-#               SIMULATION FOR A SINGLE SOUNDING
-#######################################################################
+
+###############################################################################
+#                                                                             #
+#                             Base EM1D Survey                                #
+#                                                                             #
+###############################################################################
+
+class BaseEM1DSurvey(BaseSurvey, properties.HasProperties):
+    """
+    Base EM1D survey class
+    """
+
+    def __init__(self, source_list=None, **kwargs):
+        BaseSurvey.__init__(self, source_list, **kwargs)
+
+
+###############################################################################
+#                                                                             #
+#                             Base EM1D Simulation                            #
+#                                                                             #
+###############################################################################
 
 
 class BaseEM1DSimulation(BaseSimulation):
@@ -551,454 +558,6 @@ class BaseEM1DSimulation(BaseSimulation):
         J = self.getJ(self.model)
         JtJdiag = (np.power((utils.sdiag(1./uncert)*J), 2)).sum(axis=0)
         return JtJdiag
-
-
-
-class EM1DFMSimulation(BaseEM1DSimulation):
-    """
-    Simulation class for simulating the FEM response over a 1D layered Earth
-    for a single sounding.
-    """
-
-    def __init__(self, **kwargs):
-        BaseEM1DSimulation.__init__(self, **kwargs)
-
-
-    def project_fields(self, u, output_type='response'):
-        """
-        Project from the list of Hankel transform evaluations to the data or sensitivities.
-        Data can be real or imaginary component of: total field, secondary field or ppm.
-
-        :param list u: list containing Hankel transform outputs for each unique
-        source-receiver pair.
-        :rtype: list: list containing predicted data for each unique
-        source-receiver pair.
-        :return: predicted data or sensitivities by source-receiver
-        """
-
-        COUNT = 0
-        for ii, src in enumerate(self.survey.source_list):
-            for jj, rx in enumerate(src.receiver_list):
-
-                u_temp = u[COUNT]
-
-                if rx.component == 'real':
-                    u_temp = np.real(u_temp)
-                elif rx.component == 'imag':
-                    u_temp = np.imag(u_temp)
-                elif rx.component == 'both':
-                    u_temp_r = np.real(u_temp)
-                    u_temp_i = np.imag(u_temp)
-                    if output_type == 'sensitivity_sigma':
-                        u_temp = np.vstack((u_temp_r,u_temp_i))
-                    else:
-                        u_temp = np.r_[u_temp_r,u_temp_i]
-                else:
-                    raise Exception()
-
-                # Either total or ppm
-                if rx.field_type != "secondary":
-                    u_primary = src.PrimaryField(rx.locations, rx.use_source_receiver_offset)
-                    if rx.field_type == "ppm":
-                        k = [comp == rx.orientation for comp in ["x", "y", "z"]]
-                        u_temp = 1e6 * u_temp/u_primary[0, k]
-                    else:
-                        if rx.component == 'both':
-                            if output_type == 'sensitivity_sigma':
-                                u_temp = np.vstack((u_temp_r+u_primary,u_temp_i))
-                            else:
-                                u_temp = np.r_[u_temp_r+u_primary, u_temp_i]
-
-                        else:
-                            u_temp =+ u_primary
-
-                u[COUNT] = u_temp
-                COUNT = COUNT + 1
-
-        return u
-
-
-class EM1DTMSimulation(BaseEM1DSimulation):
-    """
-    Simulation class for simulating the TEM response over a 1D layered Earth
-    for a single sounding.
-    """
-
-
-    time_intervals_are_set = False
-    frequencies_are_set = False
-    time_filter = 'key_81_CosSin_2009'
-
-
-    def __init__(self, **kwargs):
-        BaseEM1DSimulation.__init__(self, **kwargs)
-        if self.time_filter == 'key_81_CosSin_2009':
-            self.fftfilt = filters.key_81_CosSin_2009()
-        elif self.time_filter == 'key_201_CosSin_2012':
-            self.fftfilt = filters.key_201_CosSin_2012()
-        elif self.time_filter == 'key_601_CosSin_2012':
-            self.fftfilt = filters.key_601_CosSin_2009()
-        else:
-            raise Exception()
-
-
-
-    def set_time_intervals(self):
-        """
-        Define time interval for all source-receiver pairs.
-        """
-
-        for src in self.survey.source_list:
-            if src.wave_type == "general":
-                for rx in src.receiver_list:
-
-                    if src.moment_type == "single":
-                        time = rx.times
-                        pulse_period = src.pulse_period
-                        period = src.period
-                    # Dual moment
-                    else:
-                        time = np.unique(np.r_[rx.times, rx.times_dual_moment])
-                        pulse_period = np.maximum(
-                            src.pulse_period, src.pulse_period_dual_moment
-                        )
-                        period = np.maximum(src.period, src.period_dual_moment)
-                    tmin = time[time>0.].min()
-                    if src.n_pulse == 1:
-                        tmax = time.max() + pulse_period
-                    elif src.n_pulse == 2:
-                        tmax = time.max() + pulse_period + period/2.
-                    else:
-                        raise NotImplementedError("n_pulse must be either 1 or 2")
-                    n_time = int((np.log10(tmax)-np.log10(tmin))*10+1)
-
-                    rx.time_interval = np.logspace(
-                        np.log10(tmin), np.log10(tmax), n_time
-                    )
-
-        self.time_intervals_are_set = True
-
-
-    def set_frequencies(self, pts_per_dec=-1):
-        """
-        Set frequencies required for Hankel transform computation and for accurate
-        computation of the IFFT.
-        """
-
-        # Set range of time channels
-        if self.time_intervals_are_set == False:
-            self.set_time_intervals()
-
-        for src in self.survey.source_list:
-            for rx in src.receiver_list:
-
-                if src.wave_type == "general":
-                    _, freq, ft, ftarg = check_time(
-                        rx.time_interval, -1, 'dlf',
-                        {'pts_per_dec': pts_per_dec, 'dlf': self.fftfilt}, 0
-                    )
-                elif src.wave_type == "stepoff":
-                    _, freq, ft, ftarg = check_time(
-                        rx.times, -1, 'dlf',
-                        {'pts_per_dec': pts_per_dec, 'dlf': self.fftfilt}, 0,
-                    )
-                else:
-                    raise Exception("wave_type must be either general or stepoff")
-
-                rx.frequencies = freq
-                rx.ftarg = ftarg
-
-        self.frequencies_are_set = True
-
-
-    def project_fields(self, u, output_type=None):
-        """
-        Project from the list of Hankel transform evaluations to the data or sensitivities.
-
-        :param list u: list containing Hankel transform outputs for each unique
-        source-receiver pair.
-        :rtype: list: list containing predicted data for each unique
-        source-receiver pair.
-        :return: predicted data or sensitivities by source-receiver
-        """
-
-        COUNT = 0
-        for ii, src in enumerate(self.survey.source_list):
-            for jj, rx in enumerate(src.receiver_list):
-
-                u_temp = u[COUNT]
-
-                # use low-pass filter
-                if src.use_lowpass_filter:
-                    factor = src.lowpass_filter.copy()
-                else:
-                    factor = np.ones_like(rx.frequencies, dtype=complex)
-
-                # Multiplication factors
-                if rx.component in ["b", "h"]:
-                    factor *= 1./(2j*np.pi*rx.frequencies)
-
-                if rx.component in ["b", "dbdt"]:
-                    factor *= mu_0
-
-                # For stepoff waveform
-                if src.wave_type == 'stepoff':
-
-                    # Compute EM responses
-                    if u_temp.ndim == 1:
-                        resp, _ = fourier_dlf(
-                            u_temp.flatten()*factor, rx.times, rx.frequencies, rx.ftarg
-                        )
-
-                    # Compute EM sensitivities
-                    else:
-
-                        resp = np.zeros(
-                            (rx.n_time, self.n_layer), dtype=np.float64, order='F'
-                        )
-                        # TODO: remove for loop
-                        for i in range(0, self.n_layer):
-                            resp_i, _ = fourier_dlf(
-                                u_temp[:, i]*factor, rx.times, rx.frequencies, rx.ftarg
-                            )
-                            resp[:, i] = resp_i
-
-                # For general waveform.
-                # Evaluate piecewise linear input current waveforms
-                # Using Fittermann's approach (19XX) with Gaussian Quadrature
-                elif src.wave_type == 'general':
-
-                    # Compute EM responses
-                    if u_temp.ndim == 1:
-                        resp_int, _ = fourier_dlf(
-                            u_temp.flatten()*factor, rx.time_interval, rx.frequencies, rx.ftarg
-                        )
-                        # step_func = interp1d(
-                        #     self.time_int, resp_int
-                        # )
-                        step_func = iuSpline(
-                            np.log10(rx.time_interval), resp_int
-                        )
-
-                        resp = piecewise_pulse_fast(
-                            step_func, rx.times,
-                            src.time_input_currents,
-                            src.input_currents,
-                            src.period,
-                            n_pulse=src.n_pulse
-                        )
-
-                        # Compute response for the dual moment
-                        if src.moment_type == "dual":
-                            resp_dual_moment = piecewise_pulse_fast(
-                                step_func, rx.times_dual_moment,
-                                src.time_input_currents_dual_moment,
-                                src.input_currents_dual_moment,
-                                src.period_dual_moment,
-                                n_pulse=src.n_pulse
-                            )
-                            # concatenate dual moment response
-                            # so, ordering is the first moment data
-                            # then the second moment data.
-                            resp = np.r_[resp, resp_dual_moment]
-
-                    # Compute EM sensitivities
-                    else:
-                        # if src.moment_type == "single":
-                        resp = np.zeros(
-                            (rx.n_time, self.n_layer), dtype=np.float64, order='F'
-                        )
-                        # else:
-                        #     # For dual moment
-                        #     resp = np.zeros(
-                        #         (rx.n_time, self.n_layer),
-                        #         dtype=np.float64, order='F'
-                        #     )
-
-                        # TODO: remove for loop (?)
-                        for i in range(self.n_layer):
-                            resp_int_i, _ = fourier_dlf(
-                                u_temp[:, i]*factor, rx.time_interval, rx.frequencies, rx.ftarg
-                            )
-                            # step_func = interp1d(
-                            #     self.time_int, resp_int_i
-                            # )
-
-                            step_func = iuSpline(
-                                np.log10(rx.time_interval), resp_int_i
-                            )
-
-                            resp_i = piecewise_pulse_fast(
-                                step_func, rx.times,
-                                src.time_input_currents, src.input_currents,
-                                src.period, n_pulse=src.n_pulse
-                            )
-
-                            if src.moment_type == "single":
-                                resp[:, i] = resp_i
-                            else:
-                                resp_dual_moment_i = piecewise_pulse_fast(
-                                    step_func,
-                                    rx.times_dual_moment,
-                                    src.time_input_currents_dual_moment,
-                                    src.input_currents_dual_moment,
-                                    src.period_dual_moment,
-                                    n_pulse=src.n_pulse
-                                )
-                                resp[:, i] = np.r_[resp_i, resp_dual_moment_i]
-
-                u[COUNT] = resp * (-2.0/np.pi)
-                COUNT = COUNT + 1
-
-        return u
-
-
-
-#######################################################################
-#       STITCHED 1D SIMULATION CLASS AND GLOBAL FUNCTIONS
-#######################################################################
-
-def dot(args):
-    return np.dot(args[0], args[1])
-
-
-def run_simulation_FD(args):
-    """
-    This method simulates the EM response or computes the sensitivities for
-    a single sounding. The method allows for parallelization of
-    the stitched 1D problem.
-
-    :param src: a EM1DFM source object
-    :param topo: Topographic location (x, y, z)
-    :param np.array thicknesses: np.array(N-1,) layer thicknesses for a single sounding
-    :param np.array sigma: np.array(N,) layer conductivities for a single sounding
-    :param np.array eta: np.array(N,) intrinsic chargeabilities for a single sounding
-    :param np.array tau: np.array(N,) Cole-Cole time constant for a single sounding
-    :param np.array c: np.array(N,) Cole-Cole frequency distribution constant for a single sounding
-    :param np.array chi: np.array(N,) magnetic susceptibility for a single sounding
-    :param np.array dchi: np.array(N,) DC susceptibility for magnetic viscosity for a single sounding
-    :param np.array tau1: np.array(N,) lower time-relaxation constant for magnetic viscosity for a single sounding
-    :param np.array tau2: np.array(N,) upper time-relaxation constant for magnetic viscosity for a single sounding
-    :param float h: source height for a single sounding
-    :param string output_type: "response", "sensitivity_sigma", "sensitivity_height"
-    :param bool invert_height: boolean switch for inverting for source height
-    :return: response or sensitivities
-
-    """
-
-    src, topo, thicknesses, sigma, eta, tau, c, chi, dchi, tau1, tau2, h, output_type, invert_height = args
-
-    n_layer = len(thicknesses) + 1
-    local_survey = EM1DSurveyFD([src])
-    exp_map = maps.ExpMap(nP=n_layer)
-
-    if not invert_height:
-        # Use Exponential Map
-        # This is hard-wired at the moment
-
-        sim = EM1DFMSimulation(
-            survey=local_survey, thicknesses=thicknesses,
-            sigmaMap=exp_map, eta=eta, tau=tau, c=c, chi=chi, dchi=dchi, tau1=tau1, tau2=tau2,
-            topo=topo, hankel_filter='key_101_2009'
-        )
-
-        if output_type == 'sensitivity_sigma':
-            drespdsig = sim.getJ_sigma(np.log(sigma))
-            return utils.mkvc(drespdsig * sim.sigmaDeriv)
-        else:
-            resp = sim.dpred(np.log(sigma))
-            return resp
-    else:
-
-        wires = maps.Wires(('sigma', n_layer), ('h', 1))
-        sigma_map = exp_map * wires.sigma
-
-        sim = EM1DFMSimulation(
-            survey=local_survey, thicknesses=thicknesses,
-            sigmaMap=sigma_map, hMap=wires.h, topo=topo,
-            eta=eta, tau=tau, c=c, chi=chi, dchi=dchi, tau1=tau1, tau2=tau2,
-            hankel_filter='key_101_2009'
-        )
-
-        m = np.r_[np.log(sigma), h]
-        if output_type == 'sensitivity_sigma':
-            drespdsig = sim.getJ_sigma(m)
-            return utils.mkvc(drespdsig * utils.sdiag(sigma))
-            # return utils.mkvc(drespdsig)
-        elif output_type == 'sensitivity_height':
-            drespdh = sim.getJ_height(m)
-            return utils.mkvc(drespdh)
-        else:
-            resp = sim.dpred(m)
-            return resp
-
-
-def run_simulation_TD(args):
-    """
-    This method simulates the EM response or computes the sensitivities for
-    a single sounding. The method allows for parallelization of
-    the stitched 1D problem.
-
-    :param src: a EM1DTM source object
-    :param topo: Topographic location (x, y, z)
-    :param np.array thicknesses: np.array(N-1,) layer thicknesses for a single sounding
-    :param np.array sigma: np.array(N,) layer conductivities for a single sounding
-    :param np.array eta: np.array(N,) intrinsic chargeabilities for a single sounding
-    :param np.array tau: np.array(N,) Cole-Cole time constant for a single sounding
-    :param np.array c: np.array(N,) Cole-Cole frequency distribution constant for a single sounding
-    :param np.array chi: np.array(N,) magnetic susceptibility for a single sounding
-    :param np.array dchi: np.array(N,) DC susceptibility for magnetic viscosity for a single sounding
-    :param np.array tau1: np.array(N,) lower time-relaxation constant for magnetic viscosity for a single sounding
-    :param np.array tau2: np.array(N,) upper time-relaxation constant for magnetic viscosity for a single sounding
-    :param float h: source height for a single sounding
-    :param string output_type: "response", "sensitivity_sigma", "sensitivity_height"
-    :param bool invert_height: boolean switch for inverting for source height
-    :return: response or sensitivities
-
-    """
-
-    src, topo, thicknesses, sigma, eta, tau, c, chi, dchi, tau1, tau2, h, output_type, invert_height = args
-
-    n_layer = len(thicknesses) + 1
-    local_survey = EM1DSurveyTD([src])
-    exp_map = maps.ExpMap(nP=n_layer)
-
-    if not invert_height:
-        # Use Exponential Map
-        # This is hard-wired at the moment
-        sim = EM1DTMSimulation(
-            survey=local_survey, thicknesses=thicknesses,
-            sigmaMap=exp_map, eta=eta, tau=tau, c=c, chi=chi, dchi=dchi, tau1=tau1, tau2=tau2,
-            topo=topo, hankel_filter='key_101_2009'
-        )
-
-        if output_type == 'sensitivity_sigma':
-            drespdsig = sim.getJ_sigma(np.log(sigma))
-            return utils.mkvc(drespdsig * sim.sigmaDeriv)
-        else:
-            resp = sim.dpred(np.log(sigma))
-            return resp
-    else:
-
-        wires = maps.Wires(('sigma', n_layer), ('h', 1))
-        sigma_map = exp_map * wires.sigma
-        sim = EM1DTMSimulation(
-            survey=local_survey, thicknesses=thicknesses,
-            sigmaMap=sigma_map, hMap=wires.h, topo=topo,
-            eta=eta, tau=tau, c=c, chi=chi, dchi=dchi, tau1=tau1, tau2=tau2,
-            hankel_filter='key_101_2009'
-        )
-
-        m = np.r_[np.log(sigma), h]
-        if output_type == 'sensitivity_sigma':
-            drespdsig = sim.getJ_sigma(m)
-            return utils.mkvc(drespdsig * utils.sdiag(sigma))
-        elif output_type == 'sensitivity_height':
-            drespdh = sim.getJ_height(m)
-            return utils.mkvc(drespdh)
-        else:
-            resp = sim.dpred(m)
-            return resp
 
 
 class BaseStitchedEM1DSimulation(BaseSimulation):
@@ -1579,137 +1138,3 @@ class BaseStitchedEM1DSimulation(BaseSimulation):
             if self._Jmatrix_height is not None:
                 toDelete += ['_Jmatrix_height']
         return toDelete
-
-
-class StitchedEM1DFMSimulation(BaseStitchedEM1DSimulation):
-
-    def run_simulation(self, args):
-        if self.verbose:
-            print(">> Frequency-domain")
-        return run_simulation_FD(args)
-
-    # @property
-    # def frequency(self):
-    #     return self.survey.frequency
-
-    # @property
-    # def switch_real_imag(self):
-    #     return self.survey.switch_real_imag
-
-
-class StitchedEM1DTMSimulation(BaseStitchedEM1DSimulation):
-
-    # @property
-    # def wave_type(self):
-    #     return self.survey.wave_type
-
-    # @property
-    # def input_currents(self):
-    #     return self.survey.input_currents
-
-    # @property
-    # def time_input_currents(self):
-    #     return self.survey.time_input_currents
-
-    # @property
-    # def n_pulse(self):
-    #     return self.survey.n_pulse
-
-    # @property
-    # def base_frequency(self):
-    #     return self.survey.base_frequency
-
-    # @property
-    # def time(self):
-    #     return self.survey.time
-
-    # @property
-    # def use_lowpass_filter(self):
-    #     return self.survey.use_lowpass_filter
-
-    # @property
-    # def high_cut_frequency(self):
-    #     return self.survey.high_cut_frequency
-
-    # @property
-    # def moment_type(self):
-    #     return self.survey.moment_type
-
-    # @property
-    # def time_dual_moment(self):
-    #     return self.survey.time_dual_moment
-
-    # @property
-    # def time_input_currents_dual_moment(self):
-    #     return self.survey.time_input_currents_dual_moment
-
-    # @property
-    # def input_currents_dual_moment(self):
-    #     return self.survey.input_currents_dual_moment
-
-    # @property
-    # def base_frequency_dual_moment(self):
-    #     return self.survey.base_frequency_dual_moment
-
-    def run_simulation(self, args):
-        if self.verbose:
-            print(">> Time-domain")
-        return run_simulation_TD(args)
-
-    # def forward(self, m, f=None):
-    #     self.model = m
-
-    #     if self.parallel:
-    #         pool = Pool(self.n_cpu)
-    #         # This assumes the same # of layer for each of soundings
-    #         result = pool.map(
-    #             run_simulation_TD,
-    #             [
-    #                 self.input_args(i, output_type=False) for i in range(self.n_sounding)
-    #             ]
-    #         )
-    #         pool.close()
-    #         pool.join()
-    #     else:
-    #         result = [
-    #             run_simulation_TD(self.input_args(i, output_type=False)) for i in range(self.n_sounding)
-    #         ]
-    #     return np.hstack(result)
-
-    # def getJ(self, m):
-    #     """
-    #          Compute d F / d sigma
-    #     """
-    #     if self._Jmatrix is not None:
-    #         return self._Jmatrix
-    #     if self.verbose:
-    #         print(">> Compute J")
-    #     self.model = m
-    #     if self.parallel:
-    #         pool = Pool(self.n_cpu)
-    #         self._Jmatrix = pool.map(
-    #             run_simulation_TD,
-    #             [
-    #                 self.input_args(i, output_type=True) for i in range(self.n_sounding)
-    #             ]
-    #         )
-    #         pool.close()
-    #         pool.join()
-    #         if self.parallel_jvec_jtvec is False:
-    #             self._Jmatrix = sp.block_diag(self._Jmatrix).tocsr()
-    #     else:
-    #         # _Jmatrix is block diagnoal matrix (sparse)
-    #         self._Jmatrix = sp.block_diag(
-    #             [
-    #                 run_simulation_TD(self.input_args(i, output_type=True)) for i in range(self.n_sounding)
-    #             ]
-    #         ).tocsr()
-    #     return self._Jmatrix
-
-
-
-
-
-
-
-
