--- SimPEG/electromagnetics/em1d/__init__.py
@@ -1,53 +1,0 @@
-# Surveys
-from .survey import BaseEM1DSurvey, EM1DSurveyFD, EM1DSurveyTD
-
-# Sources and receivers
-from . import sources
-from . import receivers
-
-# Simulations
-from .simulation import (
-	BaseEM1DSimulation, EM1DFMSimulation, EM1DTMSimulation,
-    BaseStitchedEM1DSimulation, StitchedEM1DFMSimulation, StitchedEM1DTMSimulation
-)
-
-# Other
-from .analytics import *
-from .waveforms import *
-from . import utils
-from . import supporting_functions
-
-# from .sources import (
-#     HarmonicMagneticDipoleSource, HarmonicHorizontalLoopSource, HarmonicLineSource,
-#     TimeDomainMagneticDipoleSource, TimeDomainHorizontalLoopSource, TimeDomainLineSource
-# )
-# from .receivers import HarmonicPointReceiver, TimeDomainPointReceiver
-# from .RTEfun import rTEfunfwd, rTEfunjac
-
-from .known_waveforms import (
-    skytem_HM_2015, skytem_LM_2015,
-    butter_lowpass_filter, butterworth_type_filter,
-    piecewise_pulse, get_geotem_wave, get_nanotem_wave,
-    get_flight_direction_from_fiducial, get_rx_locations_from_flight_direction
-)
-from .known_systems import(
-    vtem_plus,
-    skytem_hm,
-    skytem_lm,
-    geotem,
-    tempest
-)
-
-
-# from .EM1DSimulation import (
-#     get_vertical_discretization_frequency,
-#     get_vertical_discretization_time,
-#     set_mesh_1d, run_simulation_FD, run_simulation_TD
-# )
-from .regularization import (
-    LateralConstraint, get_2d_mesh
-)
-from .IO import ModelIO
-import os
-import glob
-import unittest
