--- SimPEG/electromagnetics/em1d/supporting_functions/__init__.py
@@ -1,2 +1,0 @@
-from .digital_filter import *
-from .kernels import *
