--- SimPEG/electromagnetics/em1d/utils/__init__.py
@@ -1,1 +1,0 @@
-from .utils1D import *
