+++ SimPEG/electromagnetics/frequency_domain_1d/__init__.py
@@ -1,0 +1,25 @@
+# Simulations
+from .simulation import (
+	EM1DFMSimulation, StitchedEM1DFMSimulation
+)
+
+# Surveys
+from .survey import EM1DSurveyFD
+
+# Sources and receivers
+from . import sources
+from . import receivers
+
+
+
+# Other
+# from .analytics import *
+from . import supporting_functions
+
+#from .regularization import (
+#    LateralConstraint, get_2d_mesh
+#)
+from .IO import ModelIO
+import os
+import glob
+import unittest
