+++ SimPEG/electromagnetics/frequency_domain_1d/supporting_functions/__init__.py
@@ -1,0 +1,1 @@
+from .kernels import *
