+++ SimPEG/electromagnetics/frequency_domain_1d/supporting_functions/kernels.py
