+++ SimPEG/electromagnetics/time_domain_1d/known_waveforms.py
