+++ SimPEG/electromagnetics/time_domain_1d/supporting_functions/__init__.py
@@ -1,0 +1,1 @@
+from .digital_filter import *
