+++ SimPEG/electromagnetics/time_domain_1d/supporting_functions/digital_filter.py
