+++ SimPEG/electromagnetics/time_domain_1d/waveforms.py
