+++ SimPEG/electromagnetics/utils/em1d_utils.py
