+++ tutorials/13-em1d/plot_1_fdem_fwd.py
@@ -19,8 +19,8 @@ from matplotlib import pyplot as plt
 from discretize import TensorMesh
 
 from SimPEG import maps
-import simpegEM1D as em1d
-from simpegEM1D.utils import plotLayer
+from SimPEG.electromagnetics import frequency_domain_1d as em1d
+from SimPEG.electromagnetics.utils.em1d_utils import plotLayer
 
 plt.rcParams.update({'font.size': 16})
 save_file = False
